Feature/usecase

+++ domain/src/main/java/com/hmju/domain/repository/LikeRepository.kt
@@ -6,8 +6,8 @@ package com.hmju.domain.repository
  * Created by juhongmin on 2025. 4. 13.
  */
 interface LikeRepository {
-    suspend fun fetchIds(): List<Long>
-    suspend fun add(id: Long): Boolean
-    suspend fun remove(id: Long): Boolean
+    suspend fun fetchIds(): List<Int>
+    suspend fun add(id: Int): Boolean
+    suspend fun remove(id: Int): Boolean
     suspend fun clear()
 }
